+++ src/extensions/config/exceptions/invalid-config-file.tsx
@@ -1,20 +1,15 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import React from 'react';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { Color } from 'ink';
 import chalk from 'chalk';
 import { PaperError } from '../../cli';
 
 export default class InvalidConfigFile extends PaperError {
   showDoctorMessage: boolean;
-
-  render() {
-    return <Color red>{this.message}</Color>;
-  }
-
   constructor(readonly path: string) {
     super(generateMessage(path));
     this.showDoctorMessage = true;
+  }
+
+  report() {
+    return this.message;
   }
 }